Started upgrading the camera to provide a 800x600 data stream

+++ modules/support/NUbugger/src/NUbugger.cpp
@@ -56,10 +56,11 @@ namespace modules {
         };
 
         NUbugger::NUbugger(std::unique_ptr<NUClear::Environment> environment) : Reactor(std::move(environment)) {
+            
             // Set our high water mark
-            int64_t hwm = 3;
+            int hwm = 3;
             m->pub.setsockopt(ZMQ_SNDHWM, &hwm, sizeof(hwm));
-
+            
             // Bind to port 12000
             m->pub.bind("tcp://*:12000");
